cambios vista

+++ routes/web.php
@@ -17,23 +17,21 @@ Route::group(['prefix' => 'funkos'], function () {
     Route::get('/{id}', [FunkoController::class, 'show'])->name('funkos.show');
     Route::get('/edit/{id}', [FunkoController::class, 'edit'])->name('funkos.edit')->middleware(['auth', 'admin']);
     Route::put('/edit/{id}', [FunkoController::class, 'update'])->name('funkos.update')->middleware(['auth', 'admin']);
-    Route::get('/update-image/{id}', [FunkoController::class, 'editImg'])->name('funkos.editImg')->middleware(['auth', 'admin']);
+    Route::get('/update-image/{id}', [FunkoController::class, 'editImg'])->name('funkos.editImg');
     Route::patch('/update-image/{id}', [FunkoController::class, 'updateImg'])->name('funkos.updateImg')->middleware(['auth', 'admin']);
     Route::delete('/delete/{id}', [FunkoController::class, 'destroy'])->name('funkos.destroy')->middleware(['auth', 'admin']);
 });
 
 Route::group(['prefix' => 'category'], function () {
-    Route::get('/', [CategoryController::class, 'index'])->name('category.index');
-    Route::get('/create', [CategoryController::class, 'store'])->name('category.store');
-    Route::post('/create', [CategoryController::class, 'create'])->name('category.create');
-    Route::get('/{id}', [CategoryController::class, 'show'])->name('category.show');
-    Route::get('/edit/{id}', [CategoryController::class, 'edit'])->name('category.edit');
-    Route::put('/edit/{id}', [CategoryController::class, 'update'])->name('category.update');
-    Route::delete('/delete/{id}', [CategoryController::class, 'destroy'])->name('category.destroy');
-    Route::put('/activate/{id}', [CategoryController::class, 'activate'])->name('category.activate');
-})->middleware(['auth', 'admin']);
-
+    Route::get('/', [CategoryController::class, 'index'])->name('category.index')->middleware(['auth', 'admin']);
+    Route::get('/create', [CategoryController::class, 'store'])->name('category.store')->middleware(['auth', 'admin']);
+    Route::post('/create', [CategoryController::class, 'create'])->name('category.create')->middleware(['auth', 'admin']);
+    Route::get('/{id}', [CategoryController::class, 'show'])->name('category.show')->middleware(['auth', 'admin']);
+    Route::get('/edit/{id}', [CategoryController::class, 'edit'])->name('category.edit')->middleware(['auth', 'admin']);
+    Route::put('/edit/{id}', [CategoryController::class, 'update'])->name('category.update')->middleware(['auth', 'admin']);
+    Route::delete('/delete/{id}', [CategoryController::class, 'destroy'])->name('category.destroy')->middleware(['auth', 'admin']);
+    Route::put('/activate/{id}', [CategoryController::class, 'activate'])->name('category.activate')->middleware(['auth', 'admin']);
+});
 
 Auth::routes();
 
-Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
